refactor: move the data initializer class in util to config
- remove util because it is a collection of directories where interests are not clearly separated
