+++ backend/core/src/test/java/com/scoinone/core/integration/controller/BuyOrderControllerTest.java
@@ -16,7 +16,7 @@ import com.scoinone.core.service.AuthService;
 import com.scoinone.core.service.BuyOrderService;
 import com.scoinone.core.service.UserService;
 import com.scoinone.core.service.VirtualAssetService;
-import com.scoinone.core.util.UserDataInitializer;
+import com.scoinone.core.config.UserDataInitializer;
 import java.math.BigDecimal;
 import java.util.Objects;
 import org.junit.jupiter.api.AfterEach;
