+++ backend/core/src/test/java/com/scoinone/core/integration/controller/CommentControllerTest.java
@@ -19,7 +19,7 @@ import com.scoinone.core.service.AuthService;
 import com.scoinone.core.service.CommentService;
 import com.scoinone.core.service.PostService;
 import com.scoinone.core.service.UserService;
-import com.scoinone.core.util.UserDataInitializer;
+import com.scoinone.core.config.UserDataInitializer;
 import java.util.Objects;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
